Made theme & template changes; #2
There were two major theme and template changes made to this app,
following the instructions from the Youtube video tutorial
(https://www.youtube.com/watch?v=xz-yixRxZN8).

The first change was to replace a paragraph with a Polymer element
called paper-fab, or floating action button
(https://elements.polymer-project.org/elements/paper-fab)

The second change was to adjust the color palette of the app by using
[Material Palette](http://www.materialpalette.com/)'s interface to
change the theme colors to see how easy it was.

The other significant part of this toolkit is the live updating.
As soon as any files are updated, the browser is updated with the
changes automatically. Nice!